requirements.txt
apiview==1.3.25
asgiref==3.3.1
cffi==1.14.5
cryptography==3.4.6
defusedxml==0.6.0
diff-match-patch==20200713
Django==3.1.7
django-cors-headers==3.7.0
django-fernet-fields==0.6
django-import-export==2.5.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
et-xmlfile==1.0.1
jdcal==1.4.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.6
pycparser==2.20
PyJWT==1.7.1
pytz==2021.1
PyYAML==5.4.1
sqlparse==0.4.1
tablib==3.0.0
xlrd==2.0.1
xlwt==1.3.0
